Abstract response composition into controller
