+++ src/controllers/summary.ts
@@ -1,20 +1,47 @@
-import { SiteSummary } from '../../client/src/shared/payloads/SummaryResponse';
+import {
+  SiteSummary,
+  SummaryResponse,
+} from '../../client/src/shared/payloads/SummaryResponse';
 import { getManager } from 'typeorm';
 import { Child, OECReport, Organization, User } from '../entity';
 
-export async function getProviderUserCount(): Promise<number> {
+export async function getSummaryResponse(): Promise<SummaryResponse> {
+  const sites = await getSiteSummaries();
+  const summaryResponse: SummaryResponse = {
+    totalProviderUsers: await getProviderUserCount(),
+    totalOrganizations: await getOrganizationCount(),
+    totalChildren: await getChildCount(),
+    siteSummaries: sites.reduce(
+      (_siteSummaries, site) => {
+        if (!!site.submissionDate) {
+          _siteSummaries.completedSites.push(site);
+        } else if (site.totalEnrollments > 0) {
+          _siteSummaries.inProgressSites.push(site);
+        } else {
+          _siteSummaries.noDataSites.push(site);
+        }
+        return _siteSummaries;
+      },
+      { completedSites: [], inProgressSites: [], noDataSites: [] }
+    ),
+  };
+
+  return summaryResponse;
+}
+
+async function getProviderUserCount(): Promise<number> {
   return getManager().count(User, { where: { isAdmin: false } });
 }
 
-export async function getOrganizationCount(): Promise<number> {
+async function getOrganizationCount(): Promise<number> {
   return getManager().count(Organization);
 }
 
-export async function getChildCount(): Promise<number> {
+async function getChildCount(): Promise<number> {
   return getManager().count(Child);
 }
 
-export async function getSiteSummaries(): Promise<SiteSummary[]> {
+async function getSiteSummaries(): Promise<SiteSummary[]> {
   return getManager().query(`
 		SELECT s.id as siteId, s.siteName, o.id as organizationId, o.providerName, r.updatedAt, count(*) as totalEnrollments
 		FROM site s
@@ -26,6 +53,6 @@ export async function getSiteSummaries(): Promise<SiteSummary[]> {
 	`);
 }
 
-export async function getOECReports(): Promise<OECReport[]> {
+async function getOECReports(): Promise<OECReport[]> {
   return getManager().find(OECReport);
 }
