+++ src/routes/summary.ts
@@ -1,6 +1,4 @@
-import { SummaryResponse } from '../../client/src/shared/payloads/SummaryResponse';
 import express from 'express';
-import { UnauthorizedError } from 'express-jwt';
 import { ForbiddenError } from '../middleware/error/errors';
 import { passAsyncError } from '../middleware/error/passAsyncError';
 import * as controller from '../controllers/summary';
@@ -15,26 +13,6 @@ summaryRouter.get(
       throw new ForbiddenError();
     }
 
-    const sites = await controller.getSiteSummaries();
-    const summaryResponse: SummaryResponse = {
-      totalProviderUsers: await controller.getProviderUserCount(),
-      totalOrganizations: await controller.getOrganizationCount(),
-      totalChildren: await controller.getChildCount(),
-      siteSummaries: sites.reduce(
-        (_siteSummaries, site) => {
-          if (!!site.submissionDate) {
-            _siteSummaries.completedSites.push(site);
-          } else if (site.totalEnrollments > 0) {
-            _siteSummaries.inProgressSites.push(site);
-          } else {
-            _siteSummaries.noDataSites.push(site);
-          }
-          return _siteSummaries;
-        },
-        { completedSites: [], inProgressSites: [], noDataSites: [] }
-      ),
-    };
-
-    res.send(summaryResponse);
+    res.send(await controller.getSummaryResponse());
   })
 );
